+++ tests/e2e/operator/multiRegion/cockroachdb_multi_region_e2e_test.go
@@ -183,6 +183,7 @@ func (r *multiRegion) TestHelmUpgrade(t *testing.T) {
 					"--reuse-values",
 					"--set", "cockroachdb.crdbCluster.podTemplate.spec.containers[0].name=cockroachdb",
 					"--set", fmt.Sprintf("cockroachdb.crdbCluster.podTemplate.spec.containers[0].resources.requests.cpu=%s", "100m"),
+					"--set-json", `cockroachdb.crdbCluster.features=["skip-under-replicated-ranges-check"]`,
 				},
 			},
 		}